requirements.txt
Flask==0.12.2
requests==2.10.0
click==6.6
itsdangerous==0.24
Jinja2==2.6
Werkzeug==0.12.2
cherrypy==11.1.0
redis==2.10.6
Flask-Session-0.3.1
